Rename module oauth_yup_oauth2 to oauth
